refactor: Move home screen to ui/home
